src/requirements.txt
dj-database-url==0.5.0
Django==1.11.6
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
Markdown==3.0.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
